Centralize upgrade path check
